Restructure implementation
